Add 'class svn' subcommand.
This command outputs (some of) the configuration files needed to manage a
class Subversion repository. The resulting files (authz, htpasswd, etc.)
should be placed in the repository's `conf` directory.

Note: so far, we only generate the authz file and a JSON file describing the
lab groups (needed by `scripts/svn-affected-users`). We will also generate
the htpasswd file soon, but not in this commit.