Save last transcript fetch time in database.
Rather than looking for students with no transcript entries, explicitly record
the last time we looked up a student's transcript.
Some students haven't completed any courses at Memorial, but checking their
transcript over and over won't make any difference.
In the new approach, we explicitly record the last time we looked up a student's
transcript and use that information when deciding whether or not to look it up
this time.