Fix (well, hack really) lab group ordering.
I'm not sure why calling .desc() on a `Student`'s `groups` doesn't do the
right thing, but it doesn't. While I figure this out, let's just convert the
groups to a Python list that we can sort in the usual way. This shouldn't be
a huge performance problem, since the number of groups per student ought
to be low!